Refactor imgviz.io.pyglet
