+++ imgviz/_io/_pyglet/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+
+from .pyglet_imshow import pyglet_imshow
+
+from .pyglet_run import pyglet_run
