+++ imgviz/_io/_pyglet/pyglet_run.py
@@ -1,0 +1,9 @@
+from .base import check_pyglet_available
+
+
+def pyglet_run():
+    # type: () -> None
+    '''Start pyglet mainloop.'''
+    pyglet = check_pyglet_available()
+
+    return pyglet.app.run()
